requirements.txt

fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.5.2
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
opencv-python-headless==4.10.0.84
tqdm==4.66.4
matplotlib==3.8.4
torch==2.2.1
torchvision==0.17.1
shapely==2.0.4
scikit-image==0.24.0
kaggle==1.6.14
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
prometheus-client==0.20.0
streamlit==1.37.1
typer==0.12.3
